pyproject.toml
[tool.poetry]
name = "vla-tensorrt"
version = "0.1.0"
description = "VLA TensorRT Quantization Project"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.0"
torchvision = "^0.17.0"
transformers = "^4.42.0"
numpy = "^1.24.0"
opencv-python = "^4.8.0"
pillow = "^10.0.0"
tensorrt = "^8.6.0"
pycuda = "^2023.1"
onnx = "^1.15.0"
onnxruntime-gpu = "^1.16.0"
h5py = "^3.14.0"
tqdm = "^4.67.1"
matplotlib = "^3.10.5"
einops = "^0.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
